requirements-dev.txt
black~=21.6b0
flake8-annotations~=2.6.2
flake8-bugbear~=21.4.3
flake8~=3.9.2
isort~=5.9.1
mypy
pytest-cov==2.12.1
pytest-django==4.4.0
pytest==6.2.4
